Remove LabellingManager component
